Create an IndexAsset and IndexPage component
Trying to get a handle on an approach to representing assets that use
other assets. This feels clunky for now, but it works.

Also tweaked layout to be able to take arbitrary metadata, starting to
think about how assets might have a representation of metadata which
will be useful for index pages who could use that metadata.